add caching for performance gains

+++ includes/class-setlist-importer.php
@@ -403,7 +403,11 @@ class Setlist_Importer {
 		}
 
 		// Clear statistics cache when show is created/updated
-		delete_transient( 'jww_all_time_song_stats' );
+		if ( function_exists( 'jww_clear_song_stats_caches' ) ) {
+			jww_clear_song_stats_caches();
+		} else {
+			delete_transient( 'jww_all_time_song_stats' );
+		}
 
 		return array(
 			'success'  => true,
